requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.2.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
python-decouple==3.6
requests==2.28.1
toml==0.10.2
tomli==2.0.1
urllib3==1.26.11
Werkzeug==2.2.1